Teams models moved to a Teams folder
