+++ src/LobAccelerator.Library/Interfaces/ITeamsManager.cs
@@ -1,4 +1,5 @@
 ﻿using LobAccelerator.Library.Models;
+using LobAccelerator.Library.Models.Teams;
 using System;
 using System.Collections.Generic;
 using System.Text;
