+++ src/LobAccelerator.Library/Managers/TeamsManager.cs
@@ -1,5 +1,6 @@
 ﻿using LobAccelerator.Library.Interfaces;
 using LobAccelerator.Library.Models;
+using LobAccelerator.Library.Models.Teams;
 using System;
 using System.Collections.Generic;
 using System.Text;
